move read_multiband to io
